requirements.txt
flask
sqlalchemy
pymysql
bcrypt
pyjwt
python-dotenv
Flask-SQLAlchemy